requirements.txt
asgiref==3.3.1
boto3==1.17.45
botocore==1.20.45
Django==3.1.7
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
